Second iteration jobtype management (unfinished)
Add single jobtype view, functioning deletion of unused jobtypes and preparations for deleting / changing jobtypes currently in use

+++ app/Http/Controllers/JobtypeController.php
@@ -7,9 +7,12 @@ use Illuminate\Http\Request;
 use Lara\Http\Requests;
 use Lara\Http\Controllers\Controller;
 use Lara\Jobtype;
+use Lara\ScheduleEntry;
+use Lara\Schedule;
 use Session;
 use Log;
 use Redirect;
+use View;
 
 class JobtypeController extends Controller
 {
@@ -21,6 +24,7 @@ class JobtypeController extends Controller
     public function index()
     {
         $jobtypes = Jobtype::orderBy('jbtyp_title', 'ASC')->paginate(25);
+
         return view('manageJobtypesView', ['jobtypes' => $jobtypes]);
     }
 
@@ -53,7 +57,17 @@ class JobtypeController extends Controller
      */
     public function show($id)
     {
-        //
+        // get selected jobtype
+        $current_jobtype = Jobtype::findOrFail($id);
+
+        // get a list of all available job types
+        $jobtypes = Jobtype::where('jbtyp_is_archived', '=', '0')
+                           ->orderBy('jbtyp_title', 'ASC')
+                           ->get();
+
+        $events = ScheduleEntry::where('jbtyp_id', '=', $id)->with('schedule.event')->paginate(25);
+
+        return View::make('jobTypeView', compact('current_jobtype', 'jobtypes', 'events'));
     }
 
     /**
@@ -87,17 +101,7 @@ class JobtypeController extends Controller
      */
     public function destroy($id)
     {
-        // Get all the data
-        $jobtype = Jobtype::findOrFail($id);
-        
-        // Check if jobtype exists
-        if ( is_null($jobtype) ) {
-            Session::put('message', "does not exist");
-            Session::put('msgType', 'danger');
-            return Redirect::back();
-        }
-
-        // Check credentials: you can only delete, if you have rights for marketing or management 
+        // Check credentials: you can only delete, if you have rights for marketing, section management or admin
         if(!Session::has('userId') 
             OR (Session::get('userGroup') != 'marketing'
                 AND Session::get('userGroup') != 'clubleitung'
@@ -108,18 +112,35 @@ class JobtypeController extends Controller
             return Redirect::back();
         }
 
-        // Log the action while we still have the data
-        Log::info('Jobtype deleted: ' . Session::get('userName') . ' (' . Session::get('userId') . ', ' 
-                 . Session::get('userGroup') . ') deleted Jobtype "' . $jobtype->jbtyp_title .  '".');
-        
-        // HERE FIND ALL PLACES AND DO CLEAN REFERENCES!
+        // Get all the data 
+        // (throws a 404 error if jobtype doesn't exist)
+        $jobtype = Jobtype::findOrFail($id);
 
-        // Now delete the jobtype
-        Jobtype::destroy($id);
+        // Before deleting, check if this job type is in use in any existing schedule
+        if (  ScheduleEntry::where('jbtyp_id', '=', $jobtype->id)->count() > 0  ) {
+            // CASE 1: job type still in use - let the user decide what to do in each case    
+            
+            // Inform the user about the redirect and go to detailed info about the job type selected
+            Session::put('message', 'Diensttyp wurde NICHT gelöscht, weil er noch im Einsatz ist. Hier kannst du es ändern.');
+            Session::put('msgType', 'warning');
+            return Redirect::action( 'JobtypeController@show', ['id' => $jobtype->id] );
+        } 
+        else 
+        {
+            // CASE 2: job type is not used anywhere and can be remove without side effects
+            
+            // Log the action while we still have the data
+            Log::info('Jobtype deleted: ' . 
+                      Session::get('userName') . ' (' . Session::get('userId') . ', ' . Session::get('userGroup') . 
+                      ') deleted "' . $jobtype->jbtyp_title .  '" (it wasn not used in any schedule).');
 
-        // show current month afterwards
-        Session::put('message', 'Diensttyp wurde erfolgreich gelöscht.');
-        Session::put('msgType', 'success');
-        return Redirect::back();
+            // Now delete the jobtype
+            Jobtype::destroy($id);
+
+            // Return to the management page
+            Session::put('message', 'Diensttyp wurde erfolgreich gelöscht.');
+            Session::put('msgType', 'success');
+            return Redirect::action( 'JobtypeController@index', ['id' => $jobtype->id] );
+        }
     }
 }
